celery/requirements.txt
celery==5.1.2
flower==1.0.0
redis==3.5.3
psycopg2-binary==2.9.1
requests